salaam-sweet-connect-main/backend/requirements.txt
flask>=3.0
flask-cors>=4.0
mysql-connector-python>=8.3
werkzeug>=3.0
python-dotenv>=1.0
requests>=2.31
